Refactor: categrize models by their own subpackages
