+++ backend/src/main/kotlin/org/conservationco/asanahire/AsanaHireApplication.kt
@@ -3,8 +3,8 @@ package org.conservationco.asanahire
 import com.asana.models.Portfolio
 import com.asana.models.Workspace
 import net.axay.simplekotlinmail.delivery.mailerBuilder
-import org.conservationco.asanahire.domain.asana.JobSource
-import org.conservationco.asanahire.domain.mail.Protocol
+import org.conservationco.asanahire.model.asana.JobSource
+import org.conservationco.asanahire.model.mail.Protocol
 import org.simplejavamail.api.mailer.Mailer
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
