+++ backend/src/main/kotlin/org/conservationco/asanahire/model/applicant/CompactApplicant.kt
@@ -1,4 +1,4 @@
-package org.conservationco.asanahire.domain
+package org.conservationco.asanahire.model.applicant
 
 import com.asana.models.Task
