+++ backend/src/main/kotlin/org/conservationco/asanahire/model/applicant/OriginalApplicant.kt
@@ -1,4 +1,4 @@
-package org.conservationco.asanahire.domain
+package org.conservationco.asanahire.model.applicant
 
 import com.asana.models.Attachment
 import org.conservationco.asana.serialization.AsanaSerializable
