+++ backend/src/main/kotlin/org/conservationco/asanahire/model/asana/ApplicantSyncPair.kt
@@ -1,7 +1,7 @@
-package org.conservationco.asanahire.domain.asana
+package org.conservationco.asanahire.model.asana
 
-import org.conservationco.asanahire.domain.InterviewApplicant
-import org.conservationco.asanahire.domain.OriginalApplicant
+import org.conservationco.asanahire.model.applicant.InterviewApplicant
+import org.conservationco.asanahire.model.applicant.OriginalApplicant
 
 data class ApplicantSyncPair(
     val originalApplicant: OriginalApplicant,
