+++ backend/src/main/kotlin/org/conservationco/asanahire/model/mail/template/TemplatedJobMessage.kt
@@ -1,6 +1,6 @@
-package org.conservationco.asanahire.domain.mail.template
+package org.conservationco.asanahire.model.mail.template
 
-import org.conservationco.asanahire.domain.mail.Address
+import org.conservationco.asanahire.model.mail.Address
 
 class TemplatedJobMessage(
     template: Template,
